bsp: milkv-duo-fsbl: add milk-v 256M & S boards support
Looks like the original repo (github.com/milkv-duo/milkv-duo-buildroot-libraries)
is not maintained anymore and all fixes are added directly to the
official SDK [0].

We don't need the whole SDK, but we would like to use the latest FSBL
with support of all boards (Duo, Duo-256M, Duo-S), so let's switch to
the repo of "duo-buildroot-sdk-v2" and fetch only fsbl.

We also have to generate memmap header (generated by the official SDK),
so let's move Python description of memory [1, 2, 3, 4] to the layer,
to make modifications easier if user would like to modify memory map.

memmap for 256M & S are taken from SDK-V2 (as recommended by OEM) and
memmap for classic Duo is taken from SDK-V1 [2].

0 - https://github.com/milkv-duo/duo-buildroot-sdk-v2
1 - https://github.com/milkv-duo/duo-buildroot-sdk-v2/blob/main/build/scripts/mmap_conv.py
2 - https://github.com/milkv-duo/duo-buildroot-sdk/blob/develop/build/boards/cv180x/cv1800b_milkv_duo_sd/memmap.py
3 - https://github.com/milkv-duo/duo-buildroot-sdk-v2/blob/main/build/boards/cv181x/sg2000_milkv_duos_musl_riscv64_sd/memmap.py
4 - https://github.com/milkv-duo/duo-buildroot-sdk-v2/blob/main/build/boards/cv181x/sg2002_milkv_duo256m_musl_riscv64_sd/memmap.py

Signed-off-by: GP Orcullo <kinsamanka@gmail.com>
Signed-off-by: Andrei Lalaev <andrey.lalaev@gmail.com>